Critters fix
You’re hitting two separate issues:

Build crash: Cannot find module 'critters'
That happens because experimental.optimizeCss is on in your build and Next loads the optional critters package. If it isn’t installed, the build dies when prerendering /404. You can fix this either by disabling the feature or by installing the dependency.

A couple of remaining ESLint warnings and a small useMemo pattern that can be simplified.